consumer-here/requirements.txt
kafka-python
pymongo
python-dotenv
requests
